requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Werkzeug==3.1.3
pytesseract==0.3.13
Pillow==11.1.0
APScheduler==3.10.4
gunicorn>=22.0
psycopg2-binary>=2.9
